% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterpca.R
\name{clusterpca}
\alias{clusterpca}
\title{clusterwise pca}
\usage{
clusterpca(
  X,
  clus,
  ccomp = 2,
  preproc = center(),
  colwise = TRUE,
  pcafun = NULL
)
}
\arguments{
\item{X}{the data matrix}

\item{clus}{the cluster index vector}

\item{ccomp}{the number of components to extract from each block (can vary by block or be fixed, see details).}

\item{preproc}{the preprocessor}

\item{colwise}{clusters are columnwise if \code{TRUE} otherwise they are rowwise}

\item{pcafun}{an optional function that computes PCA fit that returns a \code{bi_projector} instance. see details.}
}
\description{
Fit a separate PCA model for each cluster or block of data.
}
\examples{

X <- matrix(rnorm(10*100), 10, 100)
clus <- rep(1:10, each=10)


pres <- clusterpca(X, clus, ccomp=5, colwise=TRUE)
cf <- coef(pres)

X <- matrix(rnorm(10*100), 10, 100)
clus <- rep(1:10, each=10)
pres2 <- clusterpca(X, clus, ccomp=3, preproc=center(), colwise=TRUE)

X <- matrix(rnorm(200*100), 200, 100)
clus <- rep(1:10, each=20)
pres <- clusterpca(X, clus, ccomp=5, colwise=FALSE)
}
