% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcluspca.R
\name{hcluspca}
\alias{hcluspca}
\title{Hierarchical PCA}
\usage{
hcluspca(
  X,
  hclus,
  cuts,
  est_method = c("standard", "smooth"),
  ccomp = rep(1, length(cuts)),
  spat_smooth = rep(0, length(cuts) + 1),
  cds = NULL,
  intercept = FALSE,
  svd_method = c("fast", "base", "irlba", "propack", "rsvd", "svds"),
  preproc = center()
)
}
\arguments{
\item{X}{the data matrix}

\item{ccomp}{the number of components to extract from each block (can vary by block or be fixed, see details).}

\item{preproc}{the preprocessor}
}
\description{
A multiscale PCA that defines sub-scales using a provided hierarchical clustering.
}
\examples{

grid <- expand.grid(1:10, 1:10)

## 100 images each with 50 features
X <- matrix(rnorm(100*50), 100, 50)
cuts <- c(4, 8, 16)
hclus <- dclust::dclust(grid, nstart=10)
hres1 <- hcluspca(X, hclus, cuts, est_method="standard", ccomp=c(4,1,1,1))
ncomp(hres1) == (sum(cuts) +4)

}
