% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapca.R
\name{metapca}
\alias{metapca}
\title{PCA of PCAs}
\usage{
metapca(fits, ncomp = 2, weights = NULL, combine = c("pca", "scaled", "MFA"))
}
\arguments{
\item{fits}{a list of fitted component models of type \code{bi_projector}}

\item{ncomp}{number of final components}

\item{weights}{option non-negative weighting vector -- 1 value per fit}

\item{combine}{method for weighting input pca fits

X1 <- matrix(rnorm(20\emph{10), 20, 10)
X2 <- matrix(rnorm(20}20), 20, 20)
X3 <- matrix(rnorm(20*30), 20, 30)

pc1 <- pca(X1, ncomp=10, preproc=center())
pc2 <- pca(X2, ncomp=10, preproc=center())
pc3 <- pca(X3, ncomp=10, preproc=center())

fits <- list(pc1,pc2,pc3)
pfit <- metapca(fits, ncomp=15)
ncol(scores(pfit)) == 10}
}
\description{
metapca computes a principal components analysis over several pca fits
}
