% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musubpca.R
\name{musubpca}
\alias{musubpca}
\title{subpca applied to multiple data blocks with the same variables and clusters}
\usage{
musubpca(
  X,
  clus,
  weights = NULL,
  ncomp = 2,
  inner_ccomp = 2,
  ccomp = 2,
  inner_combine = c("pca", "scaled", "MFA"),
  combine = c("pca", "scaled", "MFA"),
  preproc = center(),
  ...
)
}
\arguments{
\item{X}{the data matrix}

\item{clus}{the cluster index vector}

\item{ccomp}{the number of components to extract from each block (can vary by block or be fixed, see details).}

\item{combine}{how method to combine cluster-wise PCAs.}

\item{preproc}{the preprocessor}

\item{...}{extra args sent to \code{metapca}}
}
\description{
multiple subpca
}
\examples{

X <- multidesign::multiblock(replicate(5, matrix(rnorm(50*50), 50, 50), simplify=FALSE))
clus <- rep(1:5, length.out=50)
ret <- musubpca(X,clus)
ret <- musubpca(X,clus, inner_ccomp=.5, ccomp=.5)
}
