% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_contrast_clusterpca.R
\name{ncomp.time_contrast_clusterpca}
\alias{ncomp.time_contrast_clusterpca}
\title{Number of components per cluster or in total}
\usage{
\method{ncomp}{time_contrast_clusterpca}(x, by_cluster = FALSE, ...)
}
\description{
Number of components per cluster or in total
}
