% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subpca.R
\name{subpca}
\alias{subpca}
\title{subpca}
\usage{
subpca(
  X,
  clus,
  weights = NULL,
  ncomp = 2,
  ccomp = 2,
  combine = c("pca", "scaled", "MFA"),
  preproc = center(),
  ...
)
}
\arguments{
\item{X}{the data matrix}

\item{clus}{the cluster index vector}

\item{ncomp}{the number of components to extract from each block (can vary by block or be fixed)}

\item{combine}{how method to combine cluster-wise PCAs.}

\item{preproc}{the preprocessor}

\item{extra}{args}
}
\description{
subpca
}
\examples{

X <- matrix(rnorm(10*100), 10, 100)
clus <- rep(1:10, each=10)
ncomp=10

pres <- subpca(X, clus, ncomp=5)
}
